add docker image wrapping rayou/sops
runs a script on startup which calls `sops -d` for each file
in `ENCRYPTED_SECRETS_PATH` and stores results in corresponding 
files in `DECRYPTED_SECRETS_PATH`. The content of `SOPS_YAML` env
var is used as the `.sops.yaml` which needs to be set determine how 
to decrypt the given files.

This is intended to run in k8s init container, which the contents of
`ENCRYPTED_SECRETS_PATH` mounted from a `ConfigMap` in which values of 
`data` fields contain outputs of `sops -e`.